move metaprogramming out of interpreter
